Procedures added
descExists: verify if a given description exists.
IN(description VARCHAR(255))
OUT(bool)

editGroup: edit a given group's description and pass. If the user is
Admin and the group exists, the edit will occur.
IN(userid INT, groupid INT, description VARCHAR(255), pass VARCHAR(255))
OUT(bool)

deleteGroup: delete a given group. If the user is Admin and the group
exists, the group will be deleted.
IN(userid INT, groupid INT)
OUT(bool)